+++ docs/references.rst
@@ -1,5 +1,11 @@
 References
 ==========
 
-.. automodule:: nekoyume.models
+.. automodule:: nekoyume.block
+    :members:
+.. automodule:: nekoyume.move
+    :members:
+.. automodule:: nekoyume.node
+    :members:
+.. automodule:: nekoyume.user
     :members:
